+++ sample/src/main/java/org/opennms/integration/api/sample/health/ChainedHealthCheck.java
@@ -33,7 +33,7 @@ import java.util.Objects;
 import org.opennms.integration.api.v1.health.Context;
 import org.opennms.integration.api.v1.health.HealthCheck;
 import org.opennms.integration.api.v1.health.Response;
-import org.opennms.integration.api.v1.health.ImmutableResponse;
+import org.opennms.integration.api.v1.health.immutables.ImmutableResponse;
 import org.opennms.integration.api.v1.health.Status;
 
 /**
