+++ sample/src/main/java/org/opennms/integration/api/sample/health/ServiceExtensionHealthCheck.java
@@ -45,7 +45,7 @@ import org.opennms.integration.api.v1.detectors.DetectorClient;
 import org.opennms.integration.api.v1.health.Context;
 import org.opennms.integration.api.v1.health.HealthCheck;
 import org.opennms.integration.api.v1.health.Response;
-import org.opennms.integration.api.v1.health.ImmutableResponse;
+import org.opennms.integration.api.v1.health.immutables.ImmutableResponse;
 import org.opennms.integration.api.v1.health.Status;
 import org.opennms.integration.api.v1.model.Node;
 import org.opennms.integration.api.v1.pollers.PollerResult;
