+++ sample/src/main/java/org/opennms/integration/api/sample/health/UserDefinedLinkHealthCheck.java
@@ -36,7 +36,7 @@ import org.opennms.integration.api.v1.dao.NodeDao;
 import org.opennms.integration.api.v1.health.Context;
 import org.opennms.integration.api.v1.health.HealthCheck;
 import org.opennms.integration.api.v1.health.Response;
-import org.opennms.integration.api.v1.health.ImmutableResponse;
+import org.opennms.integration.api.v1.health.immutables.ImmutableResponse;
 import org.opennms.integration.api.v1.health.Status;
 import org.opennms.integration.api.v1.model.Node;
 import org.opennms.integration.api.v1.topology.UserDefinedLink;
